+++ bukkit/src/main/java/com/github/games647/fastlogin/bukkit/listener/protocolsupport/ProtocolSupportListener.java
@@ -67,10 +67,10 @@ public class ProtocolSupportListener implements Listener {
                     }
 
                     if (premiumUUID != null && plugin.getConfig().getBoolean("nameChangeCheck")) {
-                        profile = plugin.getCore().getStorage().loadProfile(premiumUUID);
-                        if (profile != null) {
+                        PlayerProfile uuidProfile = plugin.getCore().getStorage().loadProfile(premiumUUID);
+                        if (uuidProfile != null) {
                             plugin.getLogger().log(Level.FINER, "Player {0} changed it's username", premiumUUID);
-                            startPremiumSession(username, loginStartEvent, false, profile);
+                            startPremiumSession(username, loginStartEvent, false, uuidProfile);
                             return;
                         }
                     }